moved functions to a better place
