+++ packages/libs/eda/src/lib/core/utils/geoVariables.ts
@@ -1,5 +1,6 @@
 import { GeoConfig } from '../types/geoConfig';
 import { StudyEntity } from '../types/study';
+import { findLast } from 'lodash';
 
 /**
  * Given an entity note in the study tree, check to see if it has the following variables
@@ -44,4 +45,46 @@ export function entityToGeoConfig(
     }
   }
   return undefined;
+}
+
+/**
+ * Given the geoConfigs (one per entity with geo variables) and an entity ID,
+ * return the geoConfig/entity that is furthest from the root that is the parent of the provided entity ID.
+ *
+ * Assumes `geoConfigs` is in root to leaf order.
+ */
+export function findLeastAncestralGeoConfig(
+  geoConfigs: GeoConfig[],
+  entityId: string
+): GeoConfig {
+  return (
+    findLast(geoConfigs, ({ entity }) =>
+      entityHasChildEntityWithId(entity, entityId)
+    ) ?? geoConfigs[0]
+  );
+}
+
+// Check if the specified entity or any of its descendants has the given entityId.
+function entityHasChildEntityWithId(
+  entity: StudyEntity,
+  entityId: string
+): boolean {
+  return (
+    entity.id === entityId ||
+    (entity.children?.some((child) =>
+      entityHasChildEntityWithId(child, entityId)
+    ) ??
+      false)
+  );
+}
+
+// simple convenience function
+// defaults to root-most geoConfig for safety.
+export function getGeoConfig(
+  geoConfigs: GeoConfig[],
+  entityId?: string
+): GeoConfig {
+  return (
+    geoConfigs.find(({ entity: { id } }) => id === entityId) ?? geoConfigs[0]
+  );
 }
